Use time delta since last frame to animate clear color, print FPS
The FPS timer helped debug a problem (on Windows) where FPS appeared to
be capped while the window was focused but went very high when
unfocused.
It turns out G-Sync was enabled only when the window was focused and was
capping it to the refresh rate. When unfocused it let loose and hit up
to 6100fps.