Refactor rigging code so the skeleton is separate from the controls
This refactor was big and messy, but hopefully makes it easier to
improve the animation system with poses and animation layers.

This also moved the foot bone root to the ankle, which makes IK
solving much simpler, but required updating a lot of animation
code which was based on animating the toe position and angle.
Animations still affect the toe position, but more math is required
to determine the ankle (foot) position and rotation.

Also added rotation to COG based on lean amount. Looks way better!