Use older Vulkan SDK version 1.3.290.0
Also update VMA and VMA-Hpp to versions released with that SDK version.

Newer SDK versions caused artifacts on Mac, which I could not figure out the cause or fix. I did bisect it to find it breaks from 1.3.290.0 to 1.3.296.0. That upgrades MoltenVK from 1.2.9 to 1.2.11. I'm guessing there's a driver bug in MoltenVK, at least on Intel.